refactor: rename InspectableWebContents{Impl =>}
