+++ shell/browser/common_web_contents_delegate.cc
@@ -215,7 +215,7 @@ void CommonWebContentsDelegate::InitWithWebContents(
       web_preferences && web_preferences->IsEnabled(options::kOffscreen);
 
   // Create InspectableWebContents.
-  web_contents_.reset(InspectableWebContents::Create(
+  web_contents_.reset(new InspectableWebContents(
       web_contents, browser_context->prefs(), is_guest));
   web_contents_->SetDelegate(this);
 }
